+++ src/main/kotlin/com/hibob/feedback/dao/EmployeesDao.kt
@@ -1,5 +1,6 @@
 package com.hibob.feedback.dao
 
+import com.hibob.academy.utils.JooqTable
 import jakarta.ws.rs.BadRequestException
 import org.jooq.DSLContext
 import org.springframework.stereotype.Repository
@@ -7,6 +8,20 @@ import java.util.*
 
 @Repository
 class EmployeesDao(private val sql: DSLContext) {
+
+    class EmployeesTable(tablename: String) : JooqTable(tablename) {
+        val employeeId = createUUIDField("id")
+        val firstName = createVarcharField("first_name")
+        val lastName = createVarcharField("last_name")
+        val role = createVarcharField("role")
+        val department = createVarcharField("department")
+        val companyId = createUUIDField("company_id")
+
+        companion object {
+            val instance = EmployeesTable("employees")
+        }
+    }
+
     private val employeesTable = EmployeesTable.instance
 
     fun createEmployee(firstName: String, lastName: String, role: String, companyId: UUID, department: String): UUID{
@@ -21,7 +36,7 @@ class EmployeesDao(private val sql: DSLContext) {
             .fetchOne()!![employeesTable.employeeId]
     }
 
-    fun getDepartmentById(employeeId: UUID, companyId: UUID): String {
+    fun getDepartmentById(employeeId: UUID, companyId: UUID): String { //MIGHT DELETE LATER
         return sql.select(employeesTable.department)
             .from(employeesTable)
             .where(employeesTable.employeeId.eq(employeeId))
