+++ src/main/kotlin/com/hibob/feedback/dao/Tables.kt
@@ -17,19 +17,6 @@ class FeedbackTable(tablename: String) : JooqTable(tablename) {
     }
 }
 
-class EmployeesTable(tablename: String) : JooqTable(tablename) {
-    val employeeId = createUUIDField("id")
-    val firstName = createVarcharField("first_name")
-    val lastName = createVarcharField("last_name")
-    val role = createVarcharField("role")
-    val department = createVarcharField("department")
-    val companyId = createUUIDField("company_id")
-
-    companion object {
-        val instance = EmployeesTable("employees")
-    }
-}
-
 class CompanyTable(tablename: String) : JooqTable(tablename) {
     val companyId = createUUIDField("id")
     val name = createVarcharField("name")
